Move files to subfolder and add CMake support
This repo did not have CMake support which would be usefull for other
external projects planning on either depending on an installation or
just adding hypothesis as a submodule.

I've added CMakeLists.txt files so that the project can be exported as
an installed target with a config module, or so that the target can be
simply loaded into a project using git submodules. There is aslo a new
target name called 'uninstall' to allow the project to be removed.