Add support for Nix flakes
Downstream dependencies might require a Nix flake derivation to add the
library as a dependency.

Add a flake.nix and a flake.lock to the root directory. This means that
is another project's flake adds this repo as a input, CMake will be run
and output the correct targets to be picked up.